requirements.txt
flask
pandas
numpy
ipython
scikit-learn
scipy
spark-sklearn
flintrock
twilio
py4j
